setting pcm data size to int16_t (android required, need to change this for winrt probably)

+++ modules/core/audio/pcmdevice.h
@@ -10,7 +10,7 @@ class AudioInterface;
 class EPT_EXTERN PCMDevice
 {
 public:
-    typedef int32_t DataType;
+    typedef int16_t DataType;
 
     virtual void open(AudioInterface *audioInterface);
     virtual void close();
